requirements/base.txt
graphene-django==2.0.0